Consolidate pod template handling in pkg/podtemplate
Move pod template stuff into pkg/podtemplate, and static pod stuff from pkg/podexecutor and pkg/rke2 into pkg/staticpodexecutor

(cherry picked from commit 51a90815dbc592ac9b82287d87dea0fe9c4f7cd1)
