+++ pkg/podtemplate/podtemplate.go
@@ -1,32 +1,21 @@
-package staticpod
+package podtemplate
 
 import (
-	"bytes"
-	"crypto/md5"
 	"crypto/sha256"
 	"encoding/hex"
-	"errors"
 	"fmt"
 	"io"
 	"io/fs"
-	"io/ioutil"
 	"os"
-	"path/filepath"
 	"sort"
 	"strings"
 
-	"github.com/google/go-containerregistry/pkg/name"
-	"github.com/k3s-io/k3s/pkg/cli/cmds"
-	pkgerrors "github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/kubernetes/pkg/util/hash"
-	"sigs.k8s.io/yaml"
 )
 
 type typeVolume string
@@ -37,137 +26,6 @@ const (
 	dir              typeVolume = "dir"
 	file             typeVolume = "file"
 )
-
-type ProbeConf struct {
-	InitialDelaySeconds int32
-	TimeoutSeconds      int32
-	FailureThreshold    int32
-	PeriodSeconds       int32
-}
-
-type ProbeConfs struct {
-	Liveness  ProbeConf
-	Readiness ProbeConf
-	Startup   ProbeConf
-}
-
-type Args struct {
-	Command         string
-	Args            []string
-	Image           name.Reference
-	Dirs            []string
-	Files           []string
-	Sockets         []string
-	CISMode         bool // CIS requires that the manifest be saved with 600 permissions
-	ExcludeFiles    []string
-	HealthExec      []string
-	HealthPort      int32
-	HealthProto     string
-	HealthPath      string
-	ReadyExec       []string
-	ReadyPort       int32
-	ReadyProto      string
-	ReadyPath       string
-	CPURequest      string
-	CPULimit        string
-	MemoryRequest   string
-	MemoryLimit     string
-	ExtraMounts     []string
-	ExtraEnv        []string
-	ProbeConfs      ProbeConfs
-	SecurityContext *v1.PodSecurityContext
-	Annotations     map[string]string
-	Privileged      bool
-}
-
-// Remove cleans up the static pod manifest for the given command from the specified directory.
-// It does not actually stop or remove the static pod from the container runtime.
-func Remove(dir, command string) error {
-	manifestPath := filepath.Join(dir, command+".yaml")
-	if err := os.Remove(manifestPath); err != nil && !errors.Is(err, os.ErrNotExist) {
-		return pkgerrors.WithMessagef(err, "failed to remove %s static pod manifest", command)
-	}
-	logrus.Infof("Removed %s static pod manifest", command)
-	return nil
-}
-
-// Run writes a static pod manifest for the given command into the specified directory.
-// Note that it does not actually run the command; the kubelet is responsible for picking up
-// the manifest and creating container to run it.
-func Run(dir string, args Args) error {
-	if cmds.AgentConfig.EnableSELinux {
-		if args.SecurityContext == nil {
-			args.SecurityContext = &v1.PodSecurityContext{}
-		}
-		if args.SecurityContext.SELinuxOptions == nil {
-			args.SecurityContext.SELinuxOptions = &v1.SELinuxOptions{
-				Type: "rke2_service_t",
-			}
-		}
-	}
-	files, err := readFiles(args.Args, args.ExcludeFiles)
-	if err != nil {
-		return err
-	}
-
-	// TODO Check to make sure we aren't double mounting directories and the files in those directories
-
-	args.Files = append(args.Files, files...)
-	pod, err := pod(args)
-	if err != nil {
-		return err
-	}
-
-	manifestPath := filepath.Join(dir, args.Command+".yaml")
-
-	// We hash the completed pod manifest use that as the UID; this mimics what upstream does:
-	// https://github.com/kubernetes/kubernetes/blob/v1.24.0/pkg/kubelet/config/common.go#L58-68
-	// We manually terminate static pods with incorrect UIDs, as the kubelet cannot be relied
-	// upon to clean up the old one while the apiserver is down.
-	// See https://github.com/rancher/rke2/issues/3387 and https://github.com/rancher/rke2/issues/3725
-	hasher := md5.New()
-	hash.DeepHashObject(hasher, pod)
-	fmt.Fprintf(hasher, "file:%s", manifestPath)
-	pod.UID = types.UID(hex.EncodeToString(hasher.Sum(nil)[0:]))
-
-	b, err := yaml.Marshal(pod)
-	if err != nil {
-		return err
-	}
-	if args.CISMode {
-		return writeFile(manifestPath, b, 0600)
-	}
-	return writeFile(manifestPath, b, 0644)
-}
-
-func writeFile(dest string, content []byte, perm fs.FileMode) error {
-	name := filepath.Base(dest)
-	dir := filepath.Dir(dest)
-	if err := os.MkdirAll(dir, 0700); err != nil {
-		return err
-	}
-
-	existing, err := ioutil.ReadFile(dest)
-	if err == nil && bytes.Equal(existing, content) {
-		return nil
-	}
-
-	// Create the new manifest in a temporary directory up one level from the static pods dir and then
-	// rename it into place.  Creating manifests in the destination directory risks the Kubelet
-	// picking them up when they're partially written, or creating duplicate pods if it picks it up
-	// before the temp file is renamed over the existing file.
-	tmpdir, err := os.MkdirTemp(filepath.Dir(dir), name)
-	if err != nil {
-		return err
-	}
-	defer os.RemoveAll(tmpdir)
-
-	tmp := filepath.Join(tmpdir, name)
-	if err := os.WriteFile(tmp, content, perm); err != nil {
-		return err
-	}
-	return os.Rename(tmp, dest)
-}
 
 func hashFiles(files []string) (string, error) {
 	h := sha256.New()
@@ -186,8 +44,12 @@ func hashFiles(files []string) (string, error) {
 	return hex.EncodeToString(h.Sum(nil)), nil
 }
 
-func pod(args Args) (*v1.Pod, error) {
-	filehash, err := hashFiles(args.Files)
+func Pod(spec *Spec) (*v1.Pod, error) {
+	if spec == nil {
+		return nil, nil
+	}
+
+	filehash, err := hashFiles(spec.Files)
 	if err != nil {
 		return nil, err
 	}
@@ -198,21 +60,22 @@ func pod(args Args) (*v1.Pod, error) {
 			Kind:       "Pod",
 		},
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      args.Command,
+			Name:      spec.Command,
 			Namespace: "kube-system",
 			Labels: map[string]string{
-				"component": args.Command,
+				"component": spec.Command,
 				"tier":      "control-plane",
 			},
-			Annotations: args.Annotations,
+			Annotations: spec.Annotations,
 		},
 		Spec: v1.PodSpec{
 			Containers: []v1.Container{
 				{
-					Name:    args.Command,
-					Image:   args.Image.Name(),
-					Command: []string{args.Command},
-					Args:    args.Args,
+					Name:    spec.Command,
+					Image:   spec.Image.Name(),
+					Command: []string{spec.Command},
+					Args:    spec.Args,
+					Ports:   spec.Ports,
 					Env: []v1.EnvVar{
 						{
 							Name:  "FILE_HASH",
@@ -223,48 +86,48 @@ func pod(args Args) (*v1.Pod, error) {
 						Requests: v1.ResourceList{},
 						Limits:   v1.ResourceList{},
 					},
-					LivenessProbe:   livenessProbe(args),
-					ReadinessProbe:  readinessProbe(args),
-					StartupProbe:    startupProbe(args),
+					LivenessProbe:   spec.livenessProbe(),
+					ReadinessProbe:  spec.readinessProbe(),
+					StartupProbe:    spec.startupProbe(),
 					ImagePullPolicy: v1.PullIfNotPresent,
 					SecurityContext: &v1.SecurityContext{
-						Privileged: &args.Privileged,
+						Privileged: &spec.Privileged,
 					},
 				},
 			},
-			HostNetwork:       true,
+			HostNetwork:       spec.HostNetwork,
 			PriorityClassName: "system-cluster-critical",
-			SecurityContext:   args.SecurityContext,
+			SecurityContext:   spec.SecurityContext,
 		},
 	}
 
-	if args.CPURequest != "" {
-		if cpuRequest, err := resource.ParseQuantity(args.CPURequest); err != nil {
-			logrus.Errorf("error parsing cpu request for static pod %s: %v", args.Command, err)
+	if spec.CPURequest != "" {
+		if cpuRequest, err := resource.ParseQuantity(spec.CPURequest); err != nil {
+			logrus.Errorf("error parsing cpu request for static pod %s: %v", spec.Command, err)
 		} else {
 			p.Spec.Containers[0].Resources.Requests[v1.ResourceCPU] = cpuRequest
 		}
 	}
 
-	if args.CPULimit != "" {
-		if cpuLimit, err := resource.ParseQuantity(args.CPULimit); err != nil {
-			logrus.Errorf("error parsing cpu limit for static pod %s: %v", args.Command, err)
+	if spec.CPULimit != "" {
+		if cpuLimit, err := resource.ParseQuantity(spec.CPULimit); err != nil {
+			logrus.Errorf("error parsing cpu limit for static pod %s: %v", spec.Command, err)
 		} else {
 			p.Spec.Containers[0].Resources.Limits[v1.ResourceCPU] = cpuLimit
 		}
 	}
 
-	if args.MemoryRequest != "" {
-		if memoryRequest, err := resource.ParseQuantity(args.MemoryRequest); err != nil {
-			logrus.Errorf("error parsing memory request for static pod %s: %v", args.Command, err)
+	if spec.MemoryRequest != "" {
+		if memoryRequest, err := resource.ParseQuantity(spec.MemoryRequest); err != nil {
+			logrus.Errorf("error parsing memory request for static pod %s: %v", spec.Command, err)
 		} else {
 			p.Spec.Containers[0].Resources.Requests[v1.ResourceMemory] = memoryRequest
 		}
 	}
 
-	if args.MemoryLimit != "" {
-		if memoryLimit, err := resource.ParseQuantity(args.MemoryLimit); err != nil {
-			logrus.Errorf("error parsing memory limit for static pod %s: %v", args.Command, err)
+	if spec.MemoryLimit != "" {
+		if memoryLimit, err := resource.ParseQuantity(spec.MemoryLimit); err != nil {
+			logrus.Errorf("error parsing memory limit for static pod %s: %v", spec.Command, err)
 		} else {
 			p.Spec.Containers[0].Resources.Limits[v1.ResourceMemory] = memoryLimit
 		}
@@ -287,12 +150,12 @@ func pod(args Args) (*v1.Pod, error) {
 		}
 	}
 
-	addVolumes(p, args.Sockets, socket)
-	addVolumes(p, args.Dirs, dir)
-	addVolumes(p, args.Files, file)
+	addVolumes(p, spec.Sockets, socket)
+	addVolumes(p, spec.Dirs, dir)
+	addVolumes(p, spec.Files, file)
 
-	addExtraMounts(p, args.ExtraMounts)
-	addExtraEnv(p, args.ExtraEnv)
+	addExtraMounts(p, spec.ExtraMounts)
+	addExtraEnv(p, spec.ExtraEnv)
 
 	return p, nil
 }
@@ -337,7 +200,6 @@ func addVolumes(p *v1.Pod, src []string, volume typeVolume) {
 }
 
 func addExtraMounts(p *v1.Pod, extraMounts []string) {
-
 	for i, rawMount := range extraMounts {
 		var sourceType v1.HostPathType
 		mount := strings.Split(rawMount, ":")
@@ -414,10 +276,10 @@ func addExtraEnv(p *v1.Pod, extraEnv []string) {
 	}
 }
 
-// readFiles takes in the arguments passed to the static pod and returns a list of all files
+// ReadFiles takes in the arguments passed to the static pod and returns a list of all files
 // embedded in those arguments to be included in the pod manifest as volumes.
 // excludeFiles are not included in the returned list.
-func readFiles(args, excludeFiles []string) ([]string, error) {
+func ReadFiles(args, excludeFiles []string) ([]string, error) {
 	files := map[string]bool{}
 	excludes := map[string]bool{}
 
@@ -480,27 +342,39 @@ func kubeconfigFiles(kubeconfig string) ([]string, error) {
 
 // livenessProbe returns a Probe, using the Health values from the provided pod args,
 // and the appropriate thresholds for liveness probing.
-func livenessProbe(args Args) *v1.Probe {
-	return createProbe(args.HealthExec, args.HealthPath, args.HealthProto, args.HealthPort, args.ProbeConfs.Liveness)
+func (s *Spec) livenessProbe() *v1.Probe {
+	var host string
+	if s.HostNetwork {
+		host = "localhost"
+	}
+	return createProbe(s.HealthExec, s.HealthScheme, host, s.HealthPath, s.HealthPort, s.ProbeConfs.Liveness)
 }
 
 // readinessProbe returns a Probe, using the Ready values from the provided pod args,
 // and the appropriate thresholds for readiness probing.
-func readinessProbe(args Args) *v1.Probe {
-	return createProbe(args.ReadyExec, args.ReadyPath, args.ReadyProto, args.ReadyPort, args.ProbeConfs.Readiness)
+func (s *Spec) readinessProbe() *v1.Probe {
+	var host string
+	if s.HostNetwork {
+		host = "localhost"
+	}
+	return createProbe(s.ReadyExec, s.ReadyScheme, host, s.ReadyPath, s.ReadyPort, s.ProbeConfs.Readiness)
 }
 
-// startupProbe returns a Probe, using the Health values from the provided pod args,
+// startupProbe returns a Probe, using the Startup values from the provided pod args,
 // and the appropriate thresholds for startup probing.
-func startupProbe(args Args) *v1.Probe {
-	return createProbe(args.HealthExec, args.HealthPath, args.HealthProto, args.HealthPort, args.ProbeConfs.Startup)
+func (s *Spec) startupProbe() *v1.Probe {
+	var host string
+	if s.HostNetwork {
+		host = "localhost"
+	}
+	return createProbe(s.StartupExec, s.StartupScheme, host, s.StartupPath, s.StartupPort, s.ProbeConfs.Startup)
 }
 
 // createProbe creates a Probe using the provided configuration.
 // If command is set, an ExecAction Probe is returned.
 // If command is empty but port is set, a HTTPGetAction Probe is returned.
 // If neither is set, no Probe is returned.
-func createProbe(command []string, path, scheme string, port int32, conf ProbeConf) *v1.Probe {
+func createProbe(command []string, scheme, host, path string, port int32, conf ProbeConf) *v1.Probe {
 	probe := &v1.Probe{
 		InitialDelaySeconds: conf.InitialDelaySeconds,
 		TimeoutSeconds:      conf.TimeoutSeconds,
@@ -518,7 +392,7 @@ func createProbe(command []string, path, scheme string, port int32, conf ProbeCo
 	} else if port != 0 {
 		probe.HTTPGet = &v1.HTTPGetAction{
 			Path:   path,
-			Host:   "localhost",
+			Host:   host,
 			Scheme: v1.URIScheme(scheme),
 			Port: intstr.IntOrString{
 				IntVal: port,
@@ -528,7 +402,7 @@ func createProbe(command []string, path, scheme string, port int32, conf ProbeCo
 			probe.HTTPGet.Scheme = v1.URISchemeHTTPS
 		}
 		if probe.HTTPGet.Path == "" {
-			probe.HTTPGet.Path = "/healthz"
+			probe.HTTPGet.Path = "/livez"
 		}
 		return probe
 	}
